Fix issue that vmUuid is not set in DetachDataVolumeFromVmMsg
This should be the exact fix for 1752 and 2286 (verified in Shengyan's
environment). 2322 is also caused by this issue - since the volume is
not detached beforehand.

There is a BAT case covering this issue and failes every day.  We will
get a PASS with this fix.

for https://github.com/zstackio/issues/issues/1752
for https://github.com/zstackio/issues/issues/2286
for https://github.com/zstackio/issues/issues/2322

Signed-off-by: David Lee <live4thee@gmail.com>