Avoid connect sharedblock primary storage flooding
In previous, it will reconnect ps on host when that ps on the host
is disconnected. When there are problem with reconnect, it may flood
hundreds of reconnect request since ping host is very frequent.

This commit introduces PrimaryStorageHostStatus.Connecting, and
the status will change to connecting when reconnect msg is sending,
so ping host extension can avoid flood connect ps on host by filter
connecting status ref out.

Related: ZSTAC-16862