[BugFix: ZSTAC-40630]
Clean up the invalid endpoint notifier

In previous versions, we used "system-endpoint" as the name of the default system endpoint notifier. After version 4.0, we changed this name to "system-alarm-endpoint". In some cases, upgrading to version 4.0 and later will retain the dirty data of "system-endpoint". This patch will delete all data of the system endpoint notifier associated with the name "system-endpoint" in the database.