[BugFix: ZSTACK-42433]
Enhance local storage host capacity recalculate

Local storage will request capacity recalculate when a local storage
host is sucessfully reconnected, new added or failed to add, because
all those situations the host's capacity is updated. But it can be confirmed
that only PrimaryStorageCapacityVO needs to be refreshed as the
LocalStorageHostRefVO is already updated before sending recalculate request.

So add a flag 'skipRecalculatePrimaryStorageCapacityExtensionPoint' to skip
extensions when only want storage capacity update to date.

Signed-off-by: AlanJager <ye.zou@zstack.io>