<fix>[vip]: allow allocate ip from address pool
1. When creating a vip:
a. If the ip in the address pool is specified, but the corresponding
ipRange is not specified, allocation from it should also be allowed.
b. If the ip in the normal segment is used up, allocation from the
address pool should be allowed even if it is not specified.

2. Add a method to IpAllocateMessage to decide whether to use the
address pool when ipRange is not specified during allocation. Just
set the corresponding value in your desired scenario, such as vip.

3. The IP address of the vm should not be allocated from the address
pool, we have tested this case.

Resolves: ZSTAC-45209

Change-Id: I6878796c776b7a78676b6477696861746975636c