<fix>[storage]:  manually sync vm state once memory snapshot created
Creating a memory snapshot will suspend the vm for a while. At this
time, the host starts to synchronize the state of the vm, and the vm
state on the ui will change to the suspended state, and will not return
to normal until the next time the vm state is synchronized.

Resolves: ZSTAC-47648

Change-Id:I6b6864802564764880066e4801796c74784276