<fix>[Firewall]: sync fw ruleSetName by bootstrap-info
1. After vrouter restarts, the corresponding relationship
between firewall rules and network is wrong

2. The firewall applies rules according to the name of
the network card. When the vrouter restarts, the network
card will be renamed, causing the firewall rules to be
sent to the wrong network. Before applying the firewall
rules after the restart, change the name of the network
card recorded by the firewall to the new name.

Resolves: ZSTAC-48096

Change-Id: I6d786b7a76736a74786d786a6a66626e6c6d7379

Signed-off-by: zhangjianjun <jianjun.zhang@zstack.io>