<fix>[localstorage]: fix volume attach and distribution error
As the logic of creating vm will be marshalled when allocating primary
storage and host, there will be an exception when a vm' root volume is
appointed to a ceph ps' non-deaulf root pool, it may be allocated to
the root default pool, so we restrain its judgement rule, and add tags
to the corresponding flows. Besides, we should take the situation of
local storage is unable into consideration, and make data volume be
in the non localstorage ps when localstorage is disable.

Resolves: ZSTAC-51283

Resolves: ZSTAC-47740

Add a case CreateVmTestMarshalFlowCase.groovy

Change-Id: I677a6d79686b797963757978727a616c6c717869