<fix>[compute]: Fix cancelLongJob error
The retry mechanism is added to the cancellation of the hot migration
host. The default retry is 3 times, and the interval between each time
is 1s.

Resolves: ZSTAC-53214

Change-Id: I7174900153214677a6b74646a786c7065696b62

<fix>[compute]: Fix cancelLongJob error

Cancel the migration host task, if the result of the migration task is
canceled, the cancellation task is successful

Update internal message parameter name

Resolves: ZSTAC-53214

Change-Id: I7174900153214677a6b74646a786c7062696b62