<fix>[kvm]:Check KVM agent port availability
This commit adds a check for the availability of KVM agent port after configuring iptables.
The iptables rules are updated to allow incoming traffic on the KVM agent port,
and then a simple port check is performed to ensure that the port is open and accessible.
This check is important to ensure that the KVM virtualization environment is
properly configured and ready for use.

If the kvmagent port is blocked and the echo fails, the kvmagent
will be restarted and the command will be reissued.
However, due to the kvmagent port being blocked, the command will
fail to be issued.
The HA fencer is unable to update the heartbeat, and during the HA checker
 check, the heartbeat will not be refreshed, potentially resulting in
split-brain VMs.

Resolves: ZSTAC-56336

Change-Id: I6c6578706c6d716f7774776b65776a6f666a6263