<fix>[localstorage]: fix local storage double count capacity
The physical capacity stored in DB has a consistency issue.If
LocalStorageHostRefVO has been manually deleted in DB, when init
primary storage on host reconnected, MN will think it's the first
time to add this host (because LocalStorageHostRefVO not found).
So the new capacity from agent will be added to PrimaryStorageCapacityVO
again. To fix it, use "newAddedHost" to judge if it's the first time.

Resolves: ZSTAC-56011

Change-Id: I76627365777a77716e7a62687a776e6571767374