<fix>[storage]: fix alarm
Between the last collection of capacity
and the next collection period,
if the time reaches a new day
and capacity data has not yet been collected,
the alarm system will use the capacity data from the previous day.
On the new day, the alarm system needs to
filter out the forecast values related to the new day.

Resolves: ZSTAC-59141

Change-Id: I7579786b6f73737a686e75756a6d616878686271