<fix>[utils]: add method PathUtil.getFileLastModifiedTime
Use the inode and modification time of the file as keys for inodeMd5sums.

Because the file inode is not necessarily changed
after the zstack-store.bin is replaced.

If inode is used as the key for inodeMd5sums,
the new zstack-store.bin cannot be delivered properly.

Resolves: ZSTAC-6653
Related: ZSTAC-59604

Change-Id: I6e69666b7765637263796468697072786b667772
(cherry picked from commit efb55fa16507ee32bf26c601a6d3d8fd14b866e5)