<fix>[thread]: restore task context before executing queue task
Note about why remove context before runQueue() and restore task
context to make sure that the log at dispatch queue level would
not be confusing because of wrong api id.

runSingleFlight() keeps the same because only one task is running
at one time so that the api id could be used to find the running
task.

Resolves: ZSTAC-63302

Change-Id: I746677716d7772696d75776c716b766167706466
Signed-off-by: AlanJager <ye.zou@zstack.io>