<fix>[conf]: add trigger to group
1
Before the modification:
The APIAddSchedulerJobGroupToSchedulerTriggerMsg/APIAddSchedulerJobsToSchedulerJobGroupMsg
both load timers onto each job.

Modifications are as follows:
APIAddSchedulerJobGroupToSchedulerTriggerMsg binds the group to the timer,
and when the timer triggers at the set period, the group handles the execution of the child jobs;

2
APIAddSchedulerJobsToSchedulerJobGroupMsg only loads the jobs into the group,
and the jobs are no longer bound to a timer;

3
The SchedulerJobGroupJobRefVO has added a priority field to
indicate the job's priority within the group.

4
The APIAddSchedulerJobsToSchedulerJobGroupMsg has a new field called priorities,
which sets the priority of the loaded jobs within the group.
If no parameter is passed, the default priority will be set.

5
The priority represents the job's priority within the group;
the smaller the value, the higher the priority. The default priority value is 0.

DBImpact

Resolves: ZSV-5764

Change-Id: I77626f7076626e62796a68686e73786e76616672