<fix>[identity]: Improve quota repair to support agent's quota
When the management node starts, the identity module checks for any
missing quotas and  creates them for accounts that lack them. However,
 in some situations, quotas need to be set or unset on the agent side,
and using the default value to repair the quota is not appropriate. This
 commit enhances the quota repair mechanism to support returning a
repair value if needed, instead of setting a default value in the
 database.

- Added support to return a repair value for quotas that need agent-side
  settings.
- Ensured that the repair mechanism does not set default values in the
  database for such quotas.
- Updated the relevant methods to handle the new repair value logic.

This change ensures that quotas requiring agent-side settings are
handled correctly without defaulting to inappropriate values.

Resolves: ZSTAC-69284

Change-Id: I747663746565627a646a71646e7a64726373756a
Signed-off-by: AlanJager <ye.zou@zstack.io>