<fix>[localstorage]: set corrent installPathRecycle hostUuid
1 Remove LocalDeleteVolumeFromSrcPSFlow and
move the logic of migrating the original volume to trash into
the migrateVolume(MigrateVolumeOnLocalStorageMsg msg) method in LocalStorageBase.

2 Create the trash before modifying the hostUuid of the volume's LocalStorageResourceRefVO.

3 Add a DeleteRecycleExtensionPoint check to detect whether the installPath is still in use by the storage.
Currently, only the local storage implementation is supported.

The logic is as follows:

1 If InstallPathRecycleInventory.hostUuid is null:
Throw an error if there exists a volume whose installPath matches the installPath in the InstallPathRecycleInventory.

2 If InstallPathRecycleInventory.hostUuid is not null:
Throw an error if there exists a volume whose installPath matches the installPath in the InstallPathRecycleInventory,
and the hostUuid in the LocalStorageResourceRefVO of that volume matches the hostUuid in the InstallPathRecycleInventory.

Resolves: ZSV-7315

Change-Id: I667471796f66666e6c69796371647264786a6372