<fix>[conf]: Mark system templates with correct GPU vendor
This change is necessary because system templates need to be marked with
the correct GPU vendor for compatibility assessment checks. This commit
addresses the problem by introducing a stored procedure  to . This
procedure ensures that system model services in the  table have their
corresponding GPU vendor information correctly recorded in the  table.
The procedure is idempotent and only targets system services.

DBImpact

Resolves: ZSTAC-79649

Change-Id: I6a6f6e697a776366686664616c77697468797a79