<fix>[ovnha]: resolve ambiguous 'type' column in OVN L2 network query
This is a regression bug. The query for L2 networks associated with an OVN controller recently introduced a join with the SystemTagVO table. Since both SystemTagVO and L2NetworkVO contain a type column, the SQL query failed with an "ambiguous column" error as the source table was not specified.
This patch fixes the issue by explicitly qualifying the type column with the correct table alias in the SQL query.

APIImpact/GlobalConfigImpact/GlobalPropertyImpact/DBImpact/ZQLImpact

Resolves/Related: ZSTAC-80696
Change-Id: cc998546-bc69-4e6c-9e47-fcadf9aebc6e